Restructuring frontend and adding UI for logging in, but still debugging
