+++ frontend/Cargo.toml
@@ -10,6 +10,7 @@ serde = { version = "1.0.203", features = ["derive"] }
 serde_json = "1.0.117"
 wasm-bindgen = "0.2.92"
 wasm-bindgen-futures = "0.4.42"
+models = { path = "../models" }
 
 [dependencies.web-sys]
 version = "0.3.69"
