+++ frontend/src/components/mod.rs
@@ -1,0 +1,1 @@
+//! frontend/src/components/mod.rs
