+++ frontend/src/pages/mod.rs
@@ -1,0 +1,7 @@
+//! frontend/src/pages/mod.rs
+//! To export pages
+mod home;
+mod new_user;
+
+pub use home::*;
+pub use new_user::*;
