+++ frontend/src/router.rs
@@ -1,0 +1,23 @@
+//! frontend/src/router.rs
+//! This holds the logic for defining url routes
+use leptos::*;
+use leptos_router::{Route, Router, Routes, A};
+
+use crate::pages::*;
+
+#[component]
+pub fn AppRouter() -> impl IntoView {
+    view! {
+        <Router>
+            <Routes>
+                <Route path="/" view=HomePage/>
+                <Route path="/home" view=HomePage/>
+                <Route path="/new-user" view=CreateNewUser />
+                <Route path="/test" view=HomePage>
+                    <Route path=":id" view=|| view! { <p>"{id}"</p> } />
+                </Route>
+                <Route path="/*any" view=|| view! { <h2>"What have you done?!?"</h2> }/>
+            </Routes>
+        </Router>
+    }
+}
